+++ src/components/features/card/search-card.tsx
@@ -1,23 +1,11 @@
 import Image from 'next/image';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { CamelCaseObject } from '@/types/common.type';
-import { Tables } from '@/types/supabase.type';
-import Bookmark from './bookmark';
-
-type bookmarks = Tables<'bookmarks'>;
-type SearchCardProp = {
-  bookmarks: CamelCaseObject<
-    Pick<bookmarks, 'user_id' | 'place' | 'place_lat' | 'place_lng'>
-  >;
-  className: string;
-  initialBookmarks: boolean;
-  image: string;
-  title: string;
-};
+import { Card, CardContent, CardHeader, CardTitle } from '../../ui/card';
+import Bookmark from '../bookmark';
+import { SearchCardProp } from '@/types/card.type';
 
 /**
  * @param bookmarks user_id place place_lat place_lng 가 담긴 단일 데이터 객체
- * @param className 스타일값
+ * @param className 스타일 값 별도 지정없을경우 none
  * @param initialBookmarks 북마크가 되어있는지에 대한 boolean
  * @param title 카드의 제목
  *
